fix(course): :safety_vest: storyline: do some wrong request, and show how important of request validation

+++ src/index.ts
@@ -50,16 +50,43 @@ app.get("/course", async (req: Request<{}, Course[]>, res) => {
   return res.status(200).json(result);
 });
 
-app.get("/course/:id", async (req: Request<{ id: string }, Course>, res) => {
-  const course = await courseRepository.getById(req.params.id);
+app.get(
+  "/course/:id",
+  async (req: Request<{ id: string }, Course | string>, res) => {
+    // UUID length
+    if (req.params.id.length != 36)
+      return res.status(400).send("ID is incorrect");
 
-  return res.status(200).json(course);
-});
+    const course = await courseRepository.getById(req.params.id);
+
+    return res.status(200).json(course);
+  }
+);
 
 app.patch(
   "/course/:id",
   async (req: Request<{ id: string }, unknown, IUpdateCourseDto>, res) => {
     const { description, duration, start_time } = req.body;
+
+    // UUID length
+    if (req.params.id.length != 36)
+      return res.status(400).send("ID is incorrect");
+
+    if (isNaN(Date.parse(start_time)))
+      return res.status(400).send(`start_time is not in Date format`);
+
+    const duration_in_seconds = Number(duration);
+
+    if (isNaN(duration_in_seconds))
+      return res.status(400).send(`Duration is ${duration_in_seconds}`);
+
+    if (duration_in_seconds > COURSE_DURATION_LIMIT)
+      return res
+        .status(400)
+        .send(`Duration is more than ${COURSE_DURATION_LIMIT}`);
+
+    if (description.length === 0)
+      return res.status(400).send(`description is empty`);
 
     const result = await courseRepository.partialUpdate(req.params.id, {
       description,
@@ -72,6 +99,10 @@ app.patch(
 );
 
 app.delete("/course/:id", async (req, res) => {
+  // UUID length
+  if (req.params.id.length != 36)
+    return res.status(400).send("ID is incorrect");
+
   const result = await courseRepository.delete(req.params.id);
 
   return res.status(200).json(result);
